+++ examples/server.cc
@@ -2465,7 +2465,7 @@ int Server::on_read(Endpoint &ep) {
   msg.msg_iovlen = 1;
 
   uint8_t msg_ctrl[CMSG_SPACE(sizeof(int)) + CMSG_SPACE(sizeof(in6_pktinfo)) +
-                   CMSG_SPACE(sizeof(uint16_t))];
+                   CMSG_SPACE(sizeof(int))];
   msg.msg_control = msg_ctrl;
 
   for (; pktcnt < 10;) {
